lcrypto: check for errors in fverify and friends
In luvit we were encountering a problem where crypto.verify was leaving
an openssl error lying around and that was causing our TLS code to freak
out since it found an error in SSL_get_error().

So, make sure we check for all cases and in the case of VerifyFinal that
if the verify fails we clear the crypto errors too.